TypeScript Removed and Update to alpha-5
