+++ .eslintcache
@@ -1,0 +1,1 @@
+[{"D:\\COde\\Web Dev\\JNTUK\\jntuk-instant\\src\\index.js":"1","D:\\COde\\Web Dev\\JNTUK\\jntuk-instant\\src\\layout\\content.js":"2","D:\\COde\\Web Dev\\JNTUK\\jntuk-instant\\src\\layout\\header.js":"3","D:\\COde\\Web Dev\\JNTUK\\jntuk-instant\\src\\main.js":"4","D:\\COde\\Web Dev\\JNTUK\\jntuk-instant\\src\\about.js":"5"},{"size":617,"mtime":1610715046698,"results":"6","hashOfConfig":"7"},{"size":640,"mtime":1610715347684,"results":"8","hashOfConfig":"7"},{"size":1624,"mtime":1610715411195,"results":"9","hashOfConfig":"7"},{"size":115,"mtime":1610715077914,"results":"10","hashOfConfig":"7"},{"size":176,"mtime":1610715068928,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},"21tzle",{"filePath":"15","messages":"16","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},"D:\\COde\\Web Dev\\JNTUK\\jntuk-instant\\src\\index.js",[],["23","24"],"D:\\COde\\Web Dev\\JNTUK\\jntuk-instant\\src\\layout\\content.js",[],"D:\\COde\\Web Dev\\JNTUK\\jntuk-instant\\src\\layout\\header.js",[],"D:\\COde\\Web Dev\\JNTUK\\jntuk-instant\\src\\main.js",[],"D:\\COde\\Web Dev\\JNTUK\\jntuk-instant\\src\\about.js",[],{"ruleId":"25","replacedBy":"26"},{"ruleId":"27","replacedBy":"28"},"no-native-reassign",["29"],"no-negated-in-lhs",["30"],"no-global-assign","no-unsafe-negation"]
