+++ package.json
@@ -1,22 +1,16 @@
 {
   "name": "jntuk-instant",
-  "version": "2.0.0-alpha-4",
+  "version": "2.0.0-alpha-5",
   "private": true,
   "homepage": "https://jntuk-instant.github.io/The-React-Port/",
   "dependencies": {
-    "@testing-library/jest-dom": "^5.11.8",
+    "@testing-library/jest-dom": "^5.11.9",
     "@testing-library/react": "^11.2.3",
     "@testing-library/user-event": "^12.6.0",
-    "@types/jest": "^26.0.20",
-    "@types/node": "^12.19.12",
-    "@types/react": "^16.14.2",
-    "@types/react-dom": "^16.9.10",
-    "gh-pages": "^3.1.0",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.1",
-    "typescript": "^4.1.3",
     "web-vitals": "^0.2.4"
   },
   "scripts": {
@@ -49,6 +43,6 @@
     "printWidth": 200
   },
   "devDependencies": {
-    "@types/react-router-dom": "^5.1.7"
+    "gh-pages": "^3.1.0"
   }
 }
