--- src/layout/content.tsx
@@ -1,17 +1,0 @@
-//import React from 'react';
-import "./content.css";
-import react from "react";
-import Routes from "../routes";
-
-class Content0 extends react.Component {
-  render() {
-    return (
-      <main>
-        <Routes />
-      </main>
-    );
-  }
-}
-
-export default Content0;
-//use classes and objects
